requirements.txt
Django==4.2
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
simplejson==3.19.1
drf-yasg==1.21.5
PyJWT==2.6.0
requests-oauthlib==1.3.1
psycopg2-binary
djangorestframework-simplejwt==5.2.2
whitenoise==6.5.0